// src/models/Language.ts
export type LanguageType = {
    SELECTED_LANGUAGE: "English" | "Albanian"
    TEMPLATE: string
    DARK_MODE: string
    SIGNIN: string
    SIGNUP: string
    CONTINUE: string
    AS_GUEST: string
    SIGNOUT: string
    FIRST_NAME: string
    LAST_NAME: string
    CONFIRM_PASSWORD: string
    DOES_NOT_MATCH: string
    USERNAME: string
    PASSWORD: string
    NEW_PASSWORD: string
    IS_REQUIRED: string
    LOADING: string
    MUST_HAVE_AT_LEAST: string
    MUST_HAVE_A_MAXIMUM_OF: string
    CHARACTERS: string
    PLEASE_TYPE_A_VALID_EMAIL: string
    BACK: string
    WELCOME: string,
    THIS_IS_AN_ANIMATED_MODAL: string
    CLOSE: string
    MAIN_SCREEN: string
    PLAY: string
    STOP: string
    PRESS: string
    SWIPE: string
    RIGHT: string
    LEFT: string
    SHOW_WELCOME_POPUP: string
    SHOW_BOTTOM_SHEET: string
    SET: string
    HOME: string
    MAP: string
    PROFILE: string
    HELLO: string
    LANGUAGE: string
    THE_OFFER: string
    NIGHTS: string
    BREAKFAST: string
    DINNER: string
    AND: string
    FOOD_NOT_INCLUDED: string
    DESCRIPTION: string
    GALLERY: string
    REFRESH: string
    INVALID_EMAIL: string
    INVALID_PASSWORD: string
    EMAIL_TAKEN: string
    EDIT: string
    EDIT_PROFILE: string
    EDIT_PASSWORD: string
    PROFILE_CHANGED_SUCCESSFULLY: string
    PASSWORD_CHANGED_SUCCESSFULLY: string
    ERROR: string
    SUCCESS: string
    SAVED: string
    YOU_HAVE_NO_SAVED_OFFERS: string
    COMMENTS: string
    WRITE_A_COMMENT: string
    SEND: string
    SEE_ALL: string
    SIGN_IN_TO_COMMENT: string
    BE_FIRST_TO_COMMENT: string
    MONDAY: string
    TUESDAY: string
    WEDNESDAY: string
    THURSDAY: string
    FRIDAY: string
    SATURDAY: string
    SUNDAY: string
    JANUARY: string
    FEBRUARY: string
    MARCH: string
    APRIL: string
    MAY: string
    JUNE: string
    JULY: string
    AUGUST: string
    SEPTEMBER: string
    OCTOBER: string
    NOVEMBER: string
    DECEMBER: string
    ATTRACTIONS: string
    CREATE_OFFER: string
    CITY: string
    IMAGE_URL: string
    PRICE: string
    FROM: string
    UNTIL: string
    MONTH: string
    CATEGORY: string
    EUROPE: string
    BALKAN: string
    ASIA: string
    BEACH: string
    OTHER: string
    TRANSPORT: string
    FOOD: string
    LUNCH: string
    ALL_INCLUSIVE: string
    ADD: string
    PRESS_ADD_TO_ADD_ATTRACTIONS: string
    NAME: string
    CONFIRM: string
    CANCEL: string
    ADD_AN_ATTRACTION: string
    SELECT: string
    OFFER_CREATED_SUCCESSFULLY: string
    DELETE: string
    DO_YOU_WANT_TO_DELETE_THIS_OFFER: string
    OFFER_DELETED_SUCCESSFULLY: string
    DONE: string
    OK: string
    CONTACT: string
    MESSAGE: string
    SEND_MESSAGE: string
    MESSAGE_SENT_SUCCESSFULLY: string
    YOU_HAVE_NO_MESSAGES: string
    INFO: string
    MESSAGES: string
    YOU: string
    ANSWER: string
    THERE_IS_NO_ANSWER_YET: string
    ADDRESS: string
    PHONE: string
    WORK_SCHEDULE: string
    SOCIAL_MEDIA: string
    EMAIL: string
}